[Refactor] ScriptView 뷰 그리는 방식 수정 #8
